requirements.txt
# Web framework
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1          # CSRF protection
flask-mail==0.10.0         # Email (password reset + reminders)
flask-migrate==4.0.7       # Database migrations (Alembic)

# Auth
authlib==1.3.1
itsdangerous==2.2.0        # Signed tokens for password reset

# HTTP / OAuth
requests==2.31.0

# Database
psycopg2-binary==2.9.9     # PostgreSQL driver
python-dotenv==1.0.1

# Validation
email-validator==2.1.1

# Production server
gunicorn==22.0.0

# Optional: Stripe billing (uncomment when ready)
# stripe==9.3.0
